requirements.txt
flask
flask-cors
google-generativeai
python-dotenv
pypdf2
python-docx